created functions that parse cmd arguments

+++ src/lab1/declarations.rs
@@ -8,8 +8,8 @@ pub const PROG_NAME: usize = 0;
 pub const CONFIG_FILE: usize = 1;
 pub const WHINGE_MODE: usize = 2;
 
-pub const ERR_CMD_LINE: usize = 1;
-pub const ERR_SCRIPT_GEN: usize = 2;
+pub const ERR_CMD_LINE: u8= 1;
+pub const ERR_SCRIPT_GEN: u8= 2;
 
 use std::sync::atomic::AtomicBool;
 pub static WHINGE_ON: AtomicBool = AtomicBool::new(false);
